Implemented shooting
Basic projectile for now

+++ Memoriam/src/object/GameObject.java
@@ -1,12 +1,9 @@
 package object;
 
+import collision.*;
 import java.awt.Color;
 import java.awt.Image;
-
-import scenes.*;
 import systems.*;
-import collision.*;
-import object.*;
 
 public class GameObject {
 
@@ -49,7 +46,7 @@ public class GameObject {
         this.renderY = y;
     }
 
-    public GameObject(int x, int y, int s)
+    public GameObject(double x, double y, int s)
     {
         this.position.x = x;
         this.position.y = y;
@@ -81,8 +78,8 @@ public class GameObject {
     public void setImage(Image i) { this.image = i; }
 
     // Getters
-    public int getX() { return this.position.x; }
-    public int getY() { return this.position.y; }
+    public double getX() { return this.position.x; }
+    public double getY() { return this.position.y; }
     public Vector2 getPosition() { return this.position; }
     public int getScale() { return this.scale; }
     public Color getColor() { return this.color; }
@@ -96,7 +93,7 @@ public class GameObject {
     }
 
     // Movement (logic position)
-    public void move(int x, int y)
+    public void move(double x, double y)
     {
         this.position.x += x;
         this.position.y -= y;
